+++ CHANGES.md
@@ -1,4 +1,9 @@
 # Changelog
 
+### v0.0.2 (16.04.2020)
+- fixed "iterator not iterable" error (by edumotya)
+- support for python3.5
+- added blocking to appsrc to reduce memory consumption by queue
+
 ### v0.0.1 (18.01.2020)
 - Released stable version